guestutil/breakpoint: remove dependency on capstone
We wanted to fetch the exact bits of the instruction overwritten by
the breakpoint, hence capstone is needed to achieve this. Later it
turns out we can just fetch 15 bytes (the maximum possible length of
an x86 instruction) and put them into the emulation buffer - there is
no need to zero out anything beyond the actual size of that
instruction. It just works.